requirements.txt
python-crontab
beautifulsoup4
pymysql
